2 jsp are working good
