+++ src/main/java/by/incubator/application/servlets/ViewCarTypesServlet.java
@@ -1,3 +1,5 @@
+package by.incubator.application.servlets;
+
 import by.incubator.application.dto.VehicleTypeService;
 import by.incubator.application.infrastrucrure.core.impl.ApplicationContext;
 import by.incubator.application.main.Main;
@@ -9,13 +11,12 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.Map;
 
-@WebServlet(name = "/viewCars")
-public class ViewCarsServlet extends HttpServlet {
-    Map<Class<?>, Class<?>> interfaceToImplementation = Main.initInterfaceToImplementation();
-    ApplicationContext context = new ApplicationContext("by.incubator.application", interfaceToImplementation);
-    VehicleTypeService vehicleTypeService = context.getObject(VehicleTypeService.class);
+@WebServlet(name = "/viewTypes")
+public class ViewCarTypesServlet extends HttpServlet {
+    VehicleTypeService vehicleTypeService = new VehicleTypeService();
 
     @Override
     public void init() throws ServletException {
@@ -24,9 +25,15 @@ public class ViewCarsServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.setAttribute("cars", vehicleTypeService.getVehicles());
+        request.setAttribute("types", vehicleTypeService.getTypes());
+
+        PrintWriter pw = response.getWriter();
+
+        pw.println("<html>");
+        pw.println("<h1> Hello world </h1>");
+        pw.println("<html>");
         RequestDispatcher dispatcher = this.getServletContext()
-                .getRequestDispatcher("/jsp/viewCarsJSP.jsp");
+                .getRequestDispatcher("/jsp/viewTypesJSP.jsp");
         dispatcher.forward(request, response);
     }
 }
