+++ src/main/java/by/incubator/application/servlets/ViewCarsServlet.java
@@ -1,3 +1,5 @@
+package by.incubator.application.servlets;
+
 import by.incubator.application.dto.VehicleTypeService;
 import by.incubator.application.infrastrucrure.core.impl.ApplicationContext;
 import by.incubator.application.main.Main;
@@ -9,12 +11,13 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.util.Map;
 
-@WebServlet(name = "/viewTypes")
-public class ViewCarTypesServlet extends HttpServlet {
-    VehicleTypeService vehicleTypeService = new VehicleTypeService();
+@WebServlet(name = "/viewCars")
+public class ViewCarsServlet extends HttpServlet {
+    Map<Class<?>, Class<?>> interfaceToImplementation = Main.initInterfaceToImplementation();
+    ApplicationContext context = new ApplicationContext("by.incubator.application", interfaceToImplementation);
+    VehicleTypeService vehicleTypeService = context.getObject(VehicleTypeService.class);
 
     @Override
     public void init() throws ServletException {
@@ -23,15 +26,9 @@ public class ViewCarTypesServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.setAttribute("types", vehicleTypeService.getTypes());
-
-        PrintWriter pw = response.getWriter();
-
-        pw.println("<html>");
-        pw.println("<h1> Hello world </h1>");
-        pw.println("<html>");
+        request.setAttribute("cars", vehicleTypeService.getVehicles());
         RequestDispatcher dispatcher = this.getServletContext()
-                .getRequestDispatcher("/jsp/viewTypesJSP.jsp");
+                .getRequestDispatcher("/jsp/viewCarsJSP.jsp");
         dispatcher.forward(request, response);
     }
 }
